Fix: don't use a different "initial" timestamp for each actor
Otherwise, this leads to different "wait" states where there is no need for
it. For example, actor "main" creates actor "ping", allocates its state
parameters (count, pong). Once this actor receives a message and becomes the
same actor, it allocates itself its new state parameters (count, pong), but at a
different address (because the timestamp is different). This was useful before
to have each actor allocate address at a different place, but is not useful
anymore, because each actor lives in a different fixpoint computation, each with
its own store and kstore.